Split SnsMessageConverter into common converter for wrapped messages
